refactor: 생성 페이지에서 편집 페이지와 겹치는 부분 추출하여 컴포넌트화
